Added author and books adapters

+++ app/src/main/java/com/dehaat/goodreads/db/entity/Author.kt
@@ -9,7 +9,7 @@ import com.dehaat.goodreads.utils.GlobalConfig.DB.Author.COLUMN_NAME
 
 @Entity(tableName = GlobalConfig.DB.Author.TABLE_NAME)
 data class Author(
-    @ColumnInfo(name = GlobalConfig.DB.COLUMN_ID) @PrimaryKey(autoGenerate = true) val id: Int,
+    @ColumnInfo(name = GlobalConfig.DB.COLUMN_ID) @PrimaryKey(autoGenerate = true) val id: Long,
     @ColumnInfo(name = COLUMN_NAME) val name: String,
     @ColumnInfo(name = COLUMN_BIO) val bio: String
 )
